requirements.txt
pandas
scikit-learn==1.3.0 # Ensure this version is compatible
joblib
ibm-watson
python-dotenv
flask
gunicorn